# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Workout Tracker API",
    "version": "1.0.0",
    "description": "API documentation for the Workout Tracker application"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local Development Server"
    }
  ],
  "paths": {
...
    "/users": {
      "get": {
        "summary": "Get all users",
        "responses": {
          "200": {
            "description": "Successful retrieval of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DisplayUser"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No users found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "summary": "Create a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUser"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisplayUser"
                }
              }
            }
          },
          "400": {
            "description": "User already exists"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/users/{user_id}": {
      "get": {
        "summary": "Get user profile by user ID",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "ID of the user to retrieve profile for",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval of user profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisplayUserProfile"
                }
              }
            }
          },
          "404": {
            "description": "User or profile not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/users/{user_id}/createprofile": {
      "post": {
        "summary": "Create a user profile",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "ID of the user to create profile for",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserProfile"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User profile created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisplayUserProfile"
                }
              }
            }
          },
          "400": {
            "description": "Profile already exists for this user"
          },
          "403": {
            "description": "User does not have permission to create this profile"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/users/{user_id}/updateprofile": {
      "put": {
        "summary": "Update a user profile",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "ID of the user to update profile for",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserProfile"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User profile updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisplayUserProfile"
                }
              }
            }
          },
          "403": {
            "description": "User does not have permission to update this profile"
          },
          "404": {
            "description": "No profile found for user"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateUser": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        },
        "required": ["username", "email", "password"]
      },
      "DisplayUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          }
        },
        "required": ["id", "username", "email"]
      },
      "UserProfile": {
        "type": "object",
        "properties": {
          "first_name": {
            "type": "string"
          },
          "last_name": {
            "type": "string",
            "nullable": true
          },
          "date_of_birth": {
            "type": "string",
            "format": "date"
          },
          "height": {
            "type": "integer"
          },
          "weight": {
            "type": "integer"
          },
          "fitness_goal": {
            "type": "string",
            "nullable": true
          },
          "fitness_level": {
            "type": "string",
            "default": "beginner"
          },
          "available_time": {
            "type": "integer",
            "nullable": true
          },
          "contact_number": {
            "type": "integer"
          },
          "country_code": {
            "type": "string"
          }
        },
        "required": ["first_name", "date_of_birth", "height", "weight", "contact_number"]
      },
      "DisplayUserProfile": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "user_id": {
            "type": "integer"
          },
          "first_name": {
            "type": "string"
          },
          "last_name": {
            "type": "string",
            "nullable": true
          },
          "date_of_birth": {
            "type": "string",
            "format": "date"
          },
          "height": {
            "type": "integer"
          },
          "weight": {
            "type": "integer"
          },
          "fitness_goal": {
            "type": "string",
            "nullable": true
          },
          "fitness_level": {
            "type": "string",
            "default": "beginner"
          },
          "available_time": {
            "type": "integer",
            "nullable": true
          },
          "contact_number": {
            "type": "integer"
          }
        },
        "required": ["id", "user_id", "first_name", "date_of_birth", "height", "weight", "contact_number"]
      },
      "UpdateUserProfile": {
        "type": "object",
        "properties": {
          "first_name": {
            "type": "string",
            "nullable": true
          },
          "last_name": {
            "type": "string",
            "nullable": true
          },
          "date_of_birth": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "height": {
            "type": "integer",
            "nullable": true
          },
          "weight": {
            "type": "integer",
            "nullable": true
          },
          "fitness_goal": {
            "type": "string",
            "nullable": true
          },
          "fitness_level": {
            "type": "string",
            "nullable": true
          },
          "available_time": {
            "type": "integer",
            "nullable": true
          },
          "contact_number": {
            "type": "integer",
            "nullable": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}